+++ data/src/test/java/com/woowahan/data/datasource/BanchansRetrofitDataSourceImplTest.kt
@@ -2,14 +2,9 @@ package com.woowahan.data.datasource
 
 import com.google.gson.Gson
 import com.woowahan.data.apiservice.BestBanchanApiService
-import com.woowahan.data.apiservice.MainDishBanchanApiService
-import com.woowahan.data.apiservice.SideDishBanchanApiService
-import com.woowahan.data.apiservice.SoupDishBanchanApiService
 import com.woowahan.data.entity.BestBanchanEntity
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.assertj.core.api.Assertions.assertThat
@@ -25,24 +20,7 @@ import java.net.HttpURLConnection
 class BanchansRetrofitDataSourceImplTest{
 
     private lateinit var mockServer: MockWebServer
-    private lateinit var realRetrofit: Retrofit
     private lateinit var mockRetrofit: Retrofit
-    private val baseUrl = "https://api.codesquad.kr/onban/"
-
-    @Before
-    fun initRetrofit(){
-        realRetrofit = Retrofit.Builder().apply {
-            baseUrl(baseUrl)
-            client(OkHttpClient.Builder().apply {
-                addInterceptor(
-                    HttpLoggingInterceptor().apply {
-                        level = HttpLoggingInterceptor.Level.BODY
-                    }
-                )
-            }.build())
-            addConverterFactory(GsonConverterFactory.create())
-        }.build()
-    }
 
     @Before fun setUpMockServer(){
         mockServer = MockWebServer()
@@ -96,55 +74,5 @@ class BanchansRetrofitDataSourceImplTest{
         assertThat(actualResult).isEqualTo(expected)
     }
 
-    @Test
-    fun fetchBestBanchans_realServerRequest_isNotNull() = runTest {
-        //Given
-        val service = realRetrofit.create(BestBanchanApiService::class.java)
 
-        //When
-        val res = service.fetchBestBanchans()
-
-        //Then
-        println(res.body()?.toString())
-        assertThat(res.body()?.body).isNotNull
-    }
-
-    @Test
-    fun fetchMainDishBanchans_realServerRequest_isNotNull() = runTest {
-        //Given
-        val service = realRetrofit.create(MainDishBanchanApiService::class.java)
-
-        //WHen
-        val res = service.fetchMainDishBanchans()
-
-        //Then
-        println(res.body()?.toString())
-        assertThat(res.body()?.body).isNotNull
-    }
-
-    @Test
-    fun fetchSoupDishBanchans_realServerRequest_isNotNull() = runTest {
-        //Given
-        val service = realRetrofit.create(SoupDishBanchanApiService::class.java)
-
-        //When
-        val res = service.fetchSoupDishBanchans()
-
-        //Then
-        println(res.body()?.toString())
-        assertThat(res.body()?.body).isNotNull
-    }
-
-    @Test
-    fun fetchSideDishBanchans_realServerRequest_isNotNull() = runTest {
-        //Given
-        val service = realRetrofit.create(SideDishBanchanApiService::class.java)
-
-        //When
-        val res = service.fetchSideDishBanchans()
-
-        //Then
-        println(res.body()?.toString())
-        assertThat(res.body()?.body).isNotNull
-    }
 }
